requirements.txt
paho-mqtt
flask
pymysql
requests
zerorpc
sqlalchemy
pyyaml

pylint
pydocstring
autopep8